Update ember-try scenarios to match supported versions.
There is code in place to support versions back to beta.15, so it seems
like we should have the tests run that far back. We can always drop
support in the future.

Each ember-data version is paired with the latest bugfix of the ember
version that was current when that ember-data version was released.

Also makes it so that a failure in one of the symbolic versions
(release, beta, canary) does not make the whole build a failure.
If there’s an incompatibility between a future version of ember or
ember-data and ember-pouch, that should be fixed. But it shouldn’t
block accepting changes that do pass on all the supported versions.