test: fix store.unloadAll race condition in tests
The test labeled "creating an associated record stores
a reference to it in the parent" was failing intermittently
with the recently added ember LTS versions (e.g. 3.4)
when run with other tests under conditions "dont save hasMany"
and "sync". This problem appears to be caused by the
test logic not waiting for `store.unloadAll()` to complete.
The Ember Data API docs state that this function schedules
its work to be done during the next run loop, so this commit
wraps these calls with `Ember.run` to ensure that it finishes
before continuing.

See also:

* https://api.emberjs.com/ember-data/3.14/classes/Store/methods/unloadAll?anchor=unloadAll
* https://github.com/pouchdb-community/ember-pouch/pull/242
* https://github.com/pouchdb-community/ember-pouch/pull/243