test: fix store.unloadAll race condition in tests
The test labeled "creating an associated record stores
a reference to it in the parent" was failing intermittently
when run with other tests under conditions "dont save hasMany"
and "sync". This problem appears to be caused by the
test logic not waiting for `store.unloadAll()` to complete.
The Ember Data API docs state that this function schedules
its work to be done during the next run loop.
However, instead of adding a delay and making assumptions
about this, we wrap calls to this function in `run` to
ensure that it has finished before continuing.

See also:

* https://api.emberjs.com/ember-data/3.14/classes/Store/methods/unloadAll?anchor=unloadAll
* https://github.com/pouchdb-community/ember-pouch/pull/242
* https://github.com/pouchdb-community/ember-pouch/pull/243